+++ packages/werkbank/lib/src/addons/src/accessibility/src/_internal/semantics_inspector/semantics_inspector_tree.dart
@@ -8,6 +8,7 @@ import 'package:werkbank/src/addons/src/accessibility/src/_internal/semantics_in
 import 'package:werkbank/src/addons/src/accessibility/src/_internal/semantics_inspector/semantics_inspector.dart';
 import 'package:werkbank/src/addons/src/accessibility/src/_internal/semantics_inspector/semantics_inspector_controller.dart';
 import 'package:werkbank/src/addons/src/accessibility/src/_internal/semantics_monitor.dart';
+import 'package:werkbank/src/addons/src/accessibility/src/_internal/semantics_node_snapshot.dart';
 import 'package:werkbank/src/components/components.dart';
 import 'package:werkbank/src/theme/theme.dart';