+++ packages/werkbank/CHANGELOG.md
@@ -16,6 +16,7 @@
   - Add `AddonAccessor.isAddonActive(context, addonId)` to check if an addon is active.
 - Increase maximum `go_router` version to 17.x.x.
 - Mark release as incompatible with Flutter 3.38.x due to breaking changes in Flutter.
+- Add `indexInParent` property of the `SemanticsNode` to the semantics inspector. 
 
 # 0.17.3
 - When nodes of a tree in the UI become selected, their parents now expand, and they now scroll into view. This applies to: